moves user_project_binding controller
