+++ api/pkg/controller/user-project-binding/OWNERS
@@ -1,0 +1,10 @@
+# See the OWNERS docs: https://git.k8s.io/community/contributors/guide/owners.md
+
+approvers:
+  - team-ui
+
+reviewers:
+  - team-ui
+
+options:
+  no_parent_owners: true
